requirements.txt
aiosqlite>=0.17.0
hypothesis>=6.42.3
importlib-metadata==4.11.1
pydantic>=1.9.1
pytest==7.0.1
pytest-cov==3.0.0
pytest_asyncio>=0.18.3
sentinel==0.3.0
sqlalchemy>=1.4
strawberry-graphql>=0.95
